requirements.txt
streamlit
streamlit-folium
folium
requests
python-dateutil

